Revert "fix: Unify cloud connect and surface connection errors"
This reverts commit 506554fc10bd5a7e7dbb6f4f2e4cbb3feadbf9f0.
