+++ frontend/src/components/onboarding/CloudConnectingStep.tsx
@@ -1,12 +1,27 @@
 import { useEffect, useState, useRef, useCallback } from "react";
-import { Loader2, CheckCircle2, AlertCircle } from "lucide-react";
+import { Loader2, CheckCircle2 } from "lucide-react";
 import { useCloudStatus } from "../../hooks/useCloudStatus";
-import { connectToCloud } from "../../lib/cloudApi";
+import { getDaydreamUserId } from "../../lib/auth";
 import { persistSurveyAnswers } from "../../lib/onboardingStorage";
 import { useOnboarding } from "../../contexts/OnboardingContext";
 import { useTelemetry } from "../../contexts/TelemetryContext";
 import { CloudSurveyScreens, type SurveyAnswers } from "./CloudSurveyScreens";
 import { TelemetryDisclosure } from "./TelemetryDisclosure";
+
+/** Fire-and-forget: tell the backend to connect to the cloud relay. */
+async function activateCloudRelay() {
+  const userId = getDaydreamUserId();
+  if (!userId) return;
+  try {
+    await fetch("/api/v1/cloud/connect", {
+      method: "POST",
+      headers: { "Content-Type": "application/json" },
+      body: JSON.stringify({ user_id: userId }),
+    });
+  } catch (err) {
+    console.error("[Onboarding] Failed to auto-connect to cloud:", err);
+  }
+}
 
 interface CloudConnectingStepProps {
   onConnected: () => void;
@@ -17,8 +32,7 @@ export function CloudConnectingStep({
   onConnected,
   onBack,
 }: CloudConnectingStepProps) {
-  const { isConnected, isConnecting, connectStage, error, refresh } =
-    useCloudStatus();
+  const { isConnected, isConnecting, connectStage, refresh } = useCloudStatus();
   const { setOnboardingStyle, setSurveyAnswers } = useOnboarding();
   const {
     isDisclosed: telemetryDisclosed,
@@ -41,9 +55,7 @@ export function CloudConnectingStep({
   useEffect(() => {
     if (didConnect.current) return;
     didConnect.current = true;
-    connectToCloud()
-      .catch(e => console.error("[Onboarding] Cloud connect failed:", e))
-      .then(() => refresh());
+    activateCloudRelay().then(() => refresh());
   }, [refresh]);
 
   // Keep polling while this step is visible
@@ -200,18 +212,6 @@ export function CloudConnectingStep({
 
   // --- Survey done, waiting for cloud ---
   if (!isConnected) {
-    if (error) {
-      return (
-        <div className="flex flex-col items-center gap-4 w-full max-w-md mx-auto text-center">
-          <AlertCircle className="h-8 w-8 text-destructive" />
-          <h2 className="text-2xl font-semibold text-foreground">
-            Cloud connection failed
-          </h2>
-          <p className="text-sm text-destructive">{error}</p>
-        </div>
-      );
-    }
-
     return (
       <div className="flex flex-col items-center gap-6 w-full max-w-md mx-auto text-center">
         <h2 className="text-2xl font-semibold text-foreground">
